taskmanager/requirements.txt
Django==3.2.20
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
pytz==2023.3
